Fix annotation_ticks
Use for putting minor ticks on secondary axes
as guide_prism_minor does not work as intended
with the dup_axis function from ggplot2.

I believe the reason for this is because of how the
create_scale function in the AxisSecondary ggproto
object in ggplot2 (ggplot2/R/axis-secondary) works.

The secondary axis functions in ggplot2 seem a
bit hacky/funny so fixing them to work with
guide_prism_minor is a bit beyond me.